Move provider select beside new session button
