+++ src/client/routes/projects/workspaces/workspace-detail-header.tsx
@@ -37,41 +37,14 @@ import {
   useWorkspacePanel,
 } from '@/components/workspace';
 import { trpc } from '@/frontend/lib/trpc';
+import {
+  EXPLICIT_SESSION_PROVIDER_OPTIONS,
+  type NewSessionProviderSelection,
+  resolveProviderSelection,
+} from '@/lib/session-provider-selection';
 import { cn } from '@/lib/utils';
 
-import type {
-  NewSessionProviderSelection,
-  useSessionManagement,
-  useWorkspaceData,
-} from './use-workspace-detail';
-
-type SessionProviderValue = 'CLAUDE' | 'CODEX';
-
-const EXPLICIT_PROVIDER_OPTIONS = [
-  { value: 'CLAUDE', label: 'Claude' },
-  { value: 'CODEX', label: 'Codex' },
-] as const;
-
-function resolveProviderSelection(value: unknown): NewSessionProviderSelection {
-  if (value === 'CLAUDE' || value === 'CODEX' || value === 'WORKSPACE_DEFAULT') {
-    return value;
-  }
-  return 'WORKSPACE_DEFAULT';
-}
-
-function resolveEffectiveSessionProvider(
-  workspaceDefaultProvider: unknown,
-  userDefaultProvider: unknown
-): SessionProviderValue {
-  if (workspaceDefaultProvider === 'CLAUDE' || workspaceDefaultProvider === 'CODEX') {
-    return workspaceDefaultProvider;
-  }
-  return userDefaultProvider === 'CODEX' ? 'CODEX' : 'CLAUDE';
-}
-
-function getProviderLabel(provider: SessionProviderValue): string {
-  return provider === 'CODEX' ? 'Codex' : 'Claude';
-}
+import type { useSessionManagement, useWorkspaceData } from './use-workspace-detail';
 
 function ToggleRightPanelButton() {
   const { rightPanelVisible, toggleRightPanel } = useWorkspacePanel();
@@ -219,46 +192,6 @@ function RatchetingToggle({
   );
 }
 
-function NewSessionProviderSelect({
-  selectedProvider,
-  setSelectedProvider,
-  disabled,
-  effectiveDefaultProvider,
-}: {
-  selectedProvider: NewSessionProviderSelection;
-  setSelectedProvider: React.Dispatch<React.SetStateAction<NewSessionProviderSelection>>;
-  disabled: boolean;
-  effectiveDefaultProvider: SessionProviderValue;
-}) {
-  const triggerLabel = getProviderLabel(
-    selectedProvider === 'WORKSPACE_DEFAULT' ? effectiveDefaultProvider : selectedProvider
-  );
-
-  return (
-    <Select
-      value={selectedProvider}
-      onValueChange={(value) => {
-        setSelectedProvider(resolveProviderSelection(value));
-      }}
-      disabled={disabled}
-    >
-      <SelectTrigger className="h-8 w-[148px] text-xs">
-        <span className="truncate">{triggerLabel}</span>
-      </SelectTrigger>
-      <SelectContent>
-        <SelectItem key="workspace-default-provider" value="WORKSPACE_DEFAULT">
-          Use default ({getProviderLabel(effectiveDefaultProvider)})
-        </SelectItem>
-        {EXPLICIT_PROVIDER_OPTIONS.map((option) => (
-          <SelectItem key={option.value} value={option.value}>
-            {option.label}
-          </SelectItem>
-        ))}
-      </SelectContent>
-    </Select>
-  );
-}
-
 function WorkspaceProviderSettings({
   workspace,
   workspaceId,
@@ -330,7 +263,7 @@ function WorkspaceProviderSettings({
               </SelectTrigger>
               <SelectContent>
                 <SelectItem value="WORKSPACE_DEFAULT">Workspace Default</SelectItem>
-                {EXPLICIT_PROVIDER_OPTIONS.map((option) => (
+                {EXPLICIT_SESSION_PROVIDER_OPTIONS.map((option) => (
                   <SelectItem key={`default-${option.value}`} value={option.value}>
                     {option.label}
                   </SelectItem>
@@ -351,7 +284,7 @@ function WorkspaceProviderSettings({
               </SelectTrigger>
               <SelectContent>
                 <SelectItem value="WORKSPACE_DEFAULT">Workspace Default</SelectItem>
-                {EXPLICIT_PROVIDER_OPTIONS.map((option) => (
+                {EXPLICIT_SESSION_PROVIDER_OPTIONS.map((option) => (
                   <SelectItem key={`ratchet-${option.value}`} value={option.value}>
                     {option.label}
                   </SelectItem>
@@ -391,8 +324,6 @@ interface WorkspaceHeaderProps {
   archivePending: boolean;
   onArchiveRequest: () => void;
   handleQuickAction: ReturnType<typeof useSessionManagement>['handleQuickAction'];
-  selectedProvider: NewSessionProviderSelection;
-  setSelectedProvider: React.Dispatch<React.SetStateAction<NewSessionProviderSelection>>;
   running: boolean;
   isCreatingSession: boolean;
   hasChanges?: boolean;
@@ -407,18 +338,10 @@ export function WorkspaceHeader({
   archivePending,
   onArchiveRequest,
   handleQuickAction,
-  selectedProvider,
-  setSelectedProvider,
   running,
   isCreatingSession,
   hasChanges,
 }: WorkspaceHeaderProps) {
-  const { data: userSettings } = trpc.userSettings.get.useQuery();
-  const effectiveDefaultProvider = resolveEffectiveSessionProvider(
-    workspace.defaultSessionProvider,
-    userSettings?.defaultSessionProvider
-  );
-
   return (
     <div className="flex flex-col md:flex-row md:items-center md:justify-between gap-1 md:gap-2 px-2 py-1.5 md:px-4 md:py-2 border-b">
       <div className="flex items-center gap-2 md:gap-3 min-w-0">
@@ -434,12 +357,6 @@ export function WorkspaceHeader({
         <WorkspaceCiStatus workspace={workspace} />
       </div>
       <div className="flex items-center justify-end gap-0.5 md:gap-1 shrink-0">
-        <NewSessionProviderSelect
-          selectedProvider={selectedProvider}
-          setSelectedProvider={setSelectedProvider}
-          disabled={isCreatingSession}
-          effectiveDefaultProvider={effectiveDefaultProvider}
-        />
         <WorkspaceProviderSettings workspace={workspace} workspaceId={workspaceId} />
         <RatchetingToggle workspace={workspace} workspaceId={workspaceId} />
         <QuickActionsMenu
