+++ src/components/workspace/main-view-tab-bar.tsx
@@ -1,8 +1,17 @@
 import type { SessionStatus as DbSessionStatus } from '@factory-factory/core';
 import { Camera, FileCode, FileDiff, Plus } from 'lucide-react';
+import type { Dispatch, SetStateAction } from 'react';
 
+import { Select, SelectContent, SelectItem, SelectTrigger } from '@/components/ui/select';
 import { TabButton } from '@/components/ui/tab-button';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
+import {
+  EXPLICIT_SESSION_PROVIDER_OPTIONS,
+  getSessionProviderLabel,
+  type NewSessionProviderSelection,
+  resolveProviderSelection,
+  type SessionProviderValue,
+} from '@/lib/session-provider-selection';
 import { cn } from '@/lib/utils';
 
 import { RatchetWrenchIcon } from './ratchet-wrench-icon';
@@ -183,6 +192,9 @@ interface MainViewTabBarProps {
   disabled?: boolean;
   /** Maximum sessions allowed per workspace */
   maxSessions?: number;
+  selectedProvider: NewSessionProviderSelection;
+  setSelectedProvider: Dispatch<SetStateAction<NewSessionProviderSelection>>;
+  effectiveDefaultProvider: SessionProviderValue;
 }
 
 export function MainViewTabBar({
@@ -195,6 +207,9 @@ export function MainViewTabBar({
   onCloseSession,
   disabled,
   maxSessions,
+  selectedProvider,
+  setSelectedProvider,
+  effectiveDefaultProvider,
 }: MainViewTabBarProps) {
   const { tabs, activeTabId, selectTab, closeTab } = useWorkspacePanel();
 
@@ -205,6 +220,10 @@ export function MainViewTabBar({
   const sessionCount = sessions?.length ?? 0;
   const isAtLimit = maxSessions !== undefined && sessionCount >= maxSessions;
   const isButtonDisabled = disabled || isAtLimit;
+  const providerTriggerLabel =
+    selectedProvider === 'WORKSPACE_DEFAULT'
+      ? getSessionProviderLabel(effectiveDefaultProvider)
+      : getSessionProviderLabel(selectedProvider);
 
   return (
     <div
@@ -232,6 +251,34 @@ export function MainViewTabBar({
           />
         );
       })}
+
+      {/* Session provider selector */}
+      {onCreateSession && (
+        <Select
+          value={selectedProvider}
+          onValueChange={(value) => {
+            setSelectedProvider(resolveProviderSelection(value));
+          }}
+          disabled={isButtonDisabled}
+        >
+          <SelectTrigger
+            aria-label="New session provider"
+            className="h-6 w-[144px] text-xs shrink-0"
+          >
+            <span className="truncate">{providerTriggerLabel}</span>
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="WORKSPACE_DEFAULT">
+              Use default ({getSessionProviderLabel(effectiveDefaultProvider)})
+            </SelectItem>
+            {EXPLICIT_SESSION_PROVIDER_OPTIONS.map((option) => (
+              <SelectItem key={option.value} value={option.value}>
+                {option.label}
+              </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+      )}
 
       {/* Add session button */}
       {onCreateSession && (
