+++ src/components/workspace/workspace-content-view.stories.tsx
@@ -30,6 +30,9 @@ const meta = {
     onCloseSession: fn(),
     maxSessions: 5,
     hasWorktreePath: true,
+    selectedProvider: 'WORKSPACE_DEFAULT',
+    setSelectedProvider: fn(),
+    effectiveDefaultProvider: 'CLAUDE',
     children: <div className="p-4">Chat content</div>,
   },
 } satisfies Meta<typeof WorkspaceContentView>;
