+++ src/components/workspace/workspace-content-view.tsx
@@ -1,6 +1,10 @@
 import type { inferRouterOutputs } from '@trpc/server';
-import type { ReactNode } from 'react';
+import type { Dispatch, ReactNode, SetStateAction } from 'react';
 import type { AppRouter } from '@/frontend/lib/trpc';
+import type {
+  NewSessionProviderSelection,
+  SessionProviderValue,
+} from '@/lib/session-provider-selection';
 
 import { MainViewContent } from './main-view-content';
 import { MainViewTabBar } from './main-view-tab-bar';
@@ -28,6 +32,9 @@ interface WorkspaceContentViewProps {
   maxSessions?: number;
   /** Whether the workspace has a worktree path (required for sessions) */
   hasWorktreePath: boolean;
+  selectedProvider: NewSessionProviderSelection;
+  setSelectedProvider: Dispatch<SetStateAction<NewSessionProviderSelection>>;
+  effectiveDefaultProvider: SessionProviderValue;
 }
 
 // =============================================================================
@@ -54,6 +61,9 @@ export function WorkspaceContentView({
   children,
   maxSessions,
   hasWorktreePath,
+  selectedProvider,
+  setSelectedProvider,
+  effectiveDefaultProvider,
 }: WorkspaceContentViewProps) {
   const hasNoSessions = sessions && sessions.length === 0;
 
@@ -71,6 +81,9 @@ export function WorkspaceContentView({
           onCloseSession={onCloseSession}
           disabled={isCreatingSession || isDeletingSession || !hasWorktreePath}
           maxSessions={maxSessions}
+          selectedProvider={selectedProvider}
+          setSelectedProvider={setSelectedProvider}
+          effectiveDefaultProvider={effectiveDefaultProvider}
         />
       </div>
